Moved map load to common
