+++ src/framework/Session.cpp
@@ -342,9 +342,6 @@ void idSessionLocal::Clear()
 
 	timeHitch = 0;
 
-	rw = NULL;
-	sw = NULL;
-	menuSoundWorld = NULL;
 	readDemo = NULL;
 	writeDemo = NULL;
 	renderdemoVersion = 0;
@@ -387,10 +384,6 @@ idSessionLocal::idSessionLocal()
 {
 	guiInGame = guiMainMenu = guiIntro = guiRestartMenu = guiLoading = guiGameOver = guiActive = guiTest = guiMsg = guiMsgRestore = guiTakeNotes = NULL;
 
-	menuSoundWorld = NULL;
-
-	// lastTicMsec = 0;
-
 	Clear();
 }
 
@@ -415,18 +408,13 @@ void idSessionLocal::Stop()
 	ClearWipe();
 
 	// clear mapSpawned and demo playing flags
-	UnloadMap();
+	commonLocal.UnloadMap();
 
 	// disconnect async client
 	idAsyncNetwork::client.DisconnectFromServer();
 
 	// kill async server
 	idAsyncNetwork::server.Kill();
-
-	if (sw)
-	{
-		sw->StopAllSounds();
-	}
 
 	insideUpdateScreen = false;
 	insideExecuteMapChange = false;
@@ -450,24 +438,6 @@ void idSessionLocal::Shutdown()
 	}
 
 	Stop();
-
-	if (rw)
-	{
-		delete rw;
-		rw = NULL;
-	}
-
-	if (sw)
-	{
-		delete sw;
-		sw = NULL;
-	}
-
-	if (menuSoundWorld)
-	{
-		delete menuSoundWorld;
-		menuSoundWorld = NULL;
-	}
 
 	mapSpawnData.serverInfo.Clear();
 	mapSpawnData.syncedCVars.Clear();
@@ -915,7 +885,7 @@ void idSessionLocal::StartRecordingRenderDemo(const char *demoName)
 
 	// if we are in a map already, dump the current state
 	// sw->StartWritingDemo(writeDemo);
-	rw->StartWritingDemo(writeDemo);
+	// rw->StartWritingDemo(writeDemo);
 }
 
 /*
@@ -931,7 +901,7 @@ void idSessionLocal::StopRecordingRenderDemo()
 		return;
 	}
 	// sw->StopWritingDemo();
-	rw->StopWritingDemo();
+	// rw->StopWritingDemo();
 
 	writeDemo->Close();
 	common->Printf("stopped recording %s.\n", writeDemo->GetName());
@@ -961,8 +931,8 @@ void idSessionLocal::StopPlayingRenderDemo()
 
 	readDemo->Close();
 
-	sw->StopAllSounds();
-	soundSystem->SetPlayingSoundWorld(menuSoundWorld);
+	// sw->StopAllSounds();
+	// soundSystem->SetPlayingSoundWorld(menuSoundWorld);
 
 	common->Printf("stopped playing %s.\n", readDemo->GetName());
 	delete readDemo;
@@ -1021,10 +991,10 @@ void idSessionLocal::StartPlayingRenderDemo(idStr demoName)
 	}
 
 	// make sure localSound / GUI intro music shuts up
-	sw->StopAllSounds();
-	sw->PlayShaderDirectly("", 0);
-	menuSoundWorld->StopAllSounds();
-	menuSoundWorld->PlayShaderDirectly("", 0);
+	// sw->StopAllSounds();
+	// sw->PlayShaderDirectly("", 0);
+	// menuSoundWorld->StopAllSounds();
+	// menuSoundWorld->PlayShaderDirectly("", 0);
 
 	// exit any current game
 	Stop();
@@ -1364,7 +1334,7 @@ void idSessionLocal::MoveToNewMap(const char *mapName)
 {
 	mapSpawnData.serverInfo.Set("si_map", mapName);
 
-	ExecuteMapChange();
+	commonLocal.ExecuteMapChange();
 
 	if (!mapSpawnData.serverInfo.GetBool("devmap"))
 	{
@@ -1509,7 +1479,7 @@ void idSessionLocal::StartPlayingCmdDemo(const char *demoName)
 	LoadCmdDemoFromFile(cmdDemoFile);
 
 	// start the map
-	ExecuteMapChange();
+	commonLocal.ExecuteMapChange();
 
 	cmdDemoFile = fileSystem->OpenFileRead(fullDemoName);
 
@@ -1557,39 +1527,6 @@ void idSessionLocal::TimeCmdDemo(const char *demoName)
 	int endTime = Sys_Milliseconds();
 	sec = (endTime - startTime) / 1000.0;
 	common->Printf("%i seconds of game, replayed in %5.1f seconds\n", count / 60, sec);
-}
-
-/*
-===============
-idSessionLocal::UnloadMap
-
-Performs cleanup that needs to happen between maps, or when a
-game is exited.
-Exits with mapSpawned = false
-===============
-*/
-void idSessionLocal::UnloadMap()
-{
-	StopPlayingRenderDemo();
-
-	// end the current map in the game
-	if (game)
-	{
-		game->MapShutdown();
-	}
-
-	if (cmdDemoFile)
-	{
-		fileSystem->CloseFile(cmdDemoFile);
-		cmdDemoFile = NULL;
-	}
-
-	if (writeDemo)
-	{
-		StopRecordingRenderDemo();
-	}
-
-	mapSpawned = false;
 }
 
 /*
@@ -1677,247 +1614,6 @@ void idSessionLocal::SetBytesNeededForMapLoad(const char *mapName, int bytesNeed
 		mapDef->SetText(declText);
 		mapDef->ReplaceSourceFileText();
 	}
-}
-
-/*
-===============
-idSessionLocal::ExecuteMapChange
-
-Performs the initialization of a game based on mapSpawnData, used for both single
-player and multiplayer, but not for renderDemos, which don't
-create a game at all.
-Exits with mapSpawned = true
-===============
-*/
-void idSessionLocal::ExecuteMapChange(bool noFadeWipe)
-{
-	int i;
-	bool reloadingSameMap;
-
-	// close console and remove any prints from the notify lines
-	console->Close();
-
-	if (IsMultiplayer())
-	{
-		// make sure the mp GUI isn't up, or when players get back in the
-		// map, mpGame's menu and the gui will be out of sync.
-		SetGUI(NULL, NULL);
-	}
-
-	// mute sound
-	soundSystem->SetMute(true);
-
-	// clear all menu sounds
-	menuSoundWorld->ClearAllSoundEmitters();
-
-	// unpause the game sound world
-	// NOTE: we UnPause again later down. not sure this is needed
-	if (sw->IsPaused())
-	{
-		sw->UnPause();
-	}
-
-	// extract the map name from serverinfo
-	idStr mapString = mapSpawnData.serverInfo.GetString("si_map");
-
-	idStr fullMapName = "maps/";
-	fullMapName += mapString;
-	fullMapName.StripFileExtension();
-
-	// shut down the existing game if it is running
-	UnloadMap();
-
-	// don't do the deferred caching if we are reloading the same map
-	if (fullMapName == currentMapName)
-	{
-		reloadingSameMap = true;
-	}
-	else
-	{
-		reloadingSameMap = false;
-		currentMapName = fullMapName;
-	}
-
-	// note which media we are going to need to load
-	if (!reloadingSameMap)
-	{
-		declManager->BeginLevelLoad();
-		renderSystem->BeginLevelLoad();
-		soundSystem->BeginLevelLoad();
-	}
-
-	uiManager->BeginLevelLoad();
-	uiManager->Reload(true);
-
-	// set the loading gui that we will wipe to
-	LoadLoadingGui(mapString);
-
-	// cause prints to force screen updates as a pacifier,
-	// and draw the loading gui instead of game draws
-	insideExecuteMapChange = true;
-
-	// if this works out we will probably want all the sizes in a def file although this solution will
-	// work for new maps etc. after the first load. we can also drop the sizes into the default.cfg
-	fileSystem->ResetReadCount();
-	if (!reloadingSameMap)
-	{
-		bytesNeededForMapLoad = GetBytesNeededForMapLoad(mapString.c_str());
-	}
-	else
-	{
-		bytesNeededForMapLoad = 30 * 1024 * 1024;
-	}
-
-	ClearWipe();
-
-	// let the loading gui spin for 1 second to animate out
-	ShowLoadingGui();
-
-	// note any warning prints that happen during the load process
-	common->ClearWarnings(mapString);
-
-	// allow com_engineHz to be changed between map loads
-	com_engineHz_denominator = 100.0f * com_engineHz.GetFloat();
-	com_engineHz_latched = com_engineHz.GetFloat();
-
-	// release the mouse cursor
-	// before we do this potentially long operation
-	Sys_GrabMouseCursor(false);
-
-	// if net play, we get the number of clients during mapSpawnInfo processing
-	if (!idAsyncNetwork::IsActive())
-	{
-		numClients = 1;
-	}
-
-	int start = Sys_Milliseconds();
-
-	common->Printf("--------- Map Initialization ---------\n");
-	common->Printf("Map: %s\n", mapString.c_str());
-
-	// let the renderSystem load all the geometry
-	if (!rw->InitFromMap(fullMapName))
-	{
-		common->Error("couldn't load %s", fullMapName.c_str());
-	}
-
-	// for the synchronous networking we needed to roll the angles over from
-	// level to level, but now we can just clear everything
-	usercmdGen->InitForNewMap();
-	memset(&mapSpawnData.mapSpawnUsercmd, 0, sizeof(mapSpawnData.mapSpawnUsercmd));
-
-	// set the user info
-	for (i = 0; i < numClients; i++)
-	{
-		game->SetUserInfo(i, mapSpawnData.userInfo[i], idAsyncNetwork::client.IsActive(), false);
-		game->SetPersistentPlayerInfo(i, mapSpawnData.persistentPlayerInfo[i]);
-	}
-
-	// load and spawn all other entities ( from a savegame possibly )
-	if (loadingSaveGame && savegameFile)
-	{
-		if (game->InitFromSaveGame(fullMapName + ".map", rw, sw, savegameFile) == false)
-		{
-			// If the loadgame failed, restart the map with the player persistent data
-			loadingSaveGame = false;
-			fileSystem->CloseFile(savegameFile);
-			savegameFile = NULL;
-
-			game->SetServerInfo(mapSpawnData.serverInfo);
-			game->InitFromNewMap(fullMapName + ".map", rw, sw, idAsyncNetwork::server.IsActive(), idAsyncNetwork::client.IsActive(), Sys_Milliseconds());
-		}
-	}
-	else
-	{
-		game->SetServerInfo(mapSpawnData.serverInfo);
-		game->InitFromNewMap(fullMapName + ".map", rw, sw, idAsyncNetwork::server.IsActive(), idAsyncNetwork::client.IsActive(), Sys_Milliseconds());
-	}
-
-	if (!idAsyncNetwork::IsActive() && !loadingSaveGame)
-	{
-		// spawn players
-		for (i = 0; i < numClients; i++)
-		{
-			game->SpawnPlayer(i);
-		}
-	}
-
-	// actually purge/load the media
-	if (!reloadingSameMap)
-	{
-		renderSystem->EndLevelLoad();
-		soundSystem->EndLevelLoad();
-		declManager->EndLevelLoad();
-		SetBytesNeededForMapLoad(mapString.c_str(), fileSystem->GetReadCount());
-	}
-	uiManager->EndLevelLoad();
-
-	if (!idAsyncNetwork::IsActive() && !loadingSaveGame)
-	{
-		// run a few frames to allow everything to settle
-		for (i = 0; i < 10; i++)
-		{
-			game->RunFrame(mapSpawnData.mapSpawnUsercmd);
-		}
-	}
-
-	common->Printf("-----------------------------------\n");
-
-	int msec = Sys_Milliseconds() - start;
-	common->Printf("%6d msec to load %s\n", msec, mapString.c_str());
-
-	// let the renderSystem generate interactions now that everything is spawned
-	rw->GenerateAllInteractions();
-
-	common->PrintWarnings();
-
-	if (guiLoading && bytesNeededForMapLoad)
-	{
-		float n = fileSystem->GetReadCount();
-		float pct = (n / bytesNeededForMapLoad);
-		guiLoading->SetStateFloat("map_loading", pct);
-		guiLoading->StateChanged(com_frameTime);
-		Sys_GenerateEvents();
-		UpdateScreen();
-	}
-
-	// capture the current screen and start a wipe
-	StartWipe("wipe2Material");
-
-	usercmdGen->Clear();
-
-	// start saving commands for possible writeCmdDemo usage
-	logIndex = 0;
-	statIndex = 0;
-	lastSaveIndex = 0;
-
-	// don't bother spinning over all the tics we spent loading
-	lastGameTic = latchedTicNumber = com_ticNumber;
-
-	// remove any prints from the notify lines
-	console->ClearNotifyLines();
-
-	// stop drawing the laoding screen
-	insideExecuteMapChange = false;
-
-	Sys_SetPhysicalWorkMemory(-1, -1);
-
-	// set the game sound world for playback
-	soundSystem->SetPlayingSoundWorld(sw);
-
-	// when loading a save game the sound is paused
-	if (sw->IsPaused())
-	{
-		// unpause the game sound world
-		sw->UnPause();
-	}
-
-	// restart entity sound playback
-	soundSystem->SetMute(false);
-
-	// we are valid for game draws now
-	mapSpawned = true;
-	Sys_ClearEvents();
 }
 
 /*
@@ -2374,7 +2070,7 @@ bool idSessionLocal::LoadGame(const char *saveName)
 		// make sure no buttons are pressed
 		mapSpawnData.mapSpawnUsercmd[0].buttons = 0;
 
-		ExecuteMapChange();
+		commonLocal.ExecuteMapChange();
 
 		SetGUI(NULL, NULL);
 	}
@@ -2543,12 +2239,12 @@ void idSessionLocal::AdvanceRenderDemo(bool singleFrameOnly)
 		}
 		if (ds == DS_RENDER)
 		{
-			if (rw->ProcessDemoCommand(readDemo, &currentDemoRenderView, &demoTimeOffset))
-			{
-				// a view is ready to render
-				skipFrames--;
-				numDemoFrames++;
-			}
+			// if (rw->ProcessDemoCommand(readDemo, &currentDemoRenderView, &demoTimeOffset))
+			// {
+			// 	// a view is ready to render
+			// 	skipFrames--;
+			// 	numDemoFrames++;
+			// }
 			continue;
 		}
 		if (ds == DS_SOUND)
@@ -2693,8 +2389,8 @@ void idSessionLocal::Draw()
 	}
 	else if (readDemo)
 	{
-		rw->RenderScene(&currentDemoRenderView);
-		renderSystem->DrawDemoPics();
+		// rw->RenderScene(&currentDemoRenderView);
+		// renderSystem->DrawDemoPics();
 	}
 	else if (mapSpawned)
 	{
@@ -3008,6 +2704,7 @@ void idSessionLocal::Frame()
 		common->Printf("%i ", latchedTicNumber - lastGameTic);
 	}
 
+	// Seriously wtf is this doing why.
 	int gameTicsToRun = latchedTicNumber - lastGameTic;
 	int i;
 	for (i = 0; i < gameTicsToRun; i++)
@@ -3141,7 +2838,7 @@ void idSessionLocal::RunGameTic()
 		else if (!idStr::Icmp(args.Argv(0), "died"))
 		{
 			// restart on the same map
-			UnloadMap();
+			commonLocal.UnloadMap();
 			SetGUI(guiRestartMenu, NULL);
 		}
 		else if (!idStr::Icmp(args.Argv(0), "disconnect"))
@@ -3214,14 +2911,6 @@ void idSessionLocal::Init()
 
 	cmdSystem->AddCommand("hitch", Session_Hitch_f, CMD_FL_SYSTEM | CMD_FL_CHEAT, "hitches the game");
 
-	// the same idRenderWorld will be used for all games
-	// and demos, insuring that level specific models
-	// will be freed
-	rw = renderSystem->AllocRenderWorld();
-	sw = soundSystem->AllocSoundWorld(rw);
-
-	menuSoundWorld = soundSystem->AllocSoundWorld(rw);
-
 	// we have a single instance of the main menu
 #ifndef ID_DEMO_BUILD
 	guiMainMenu = uiManager->FindGui("guis/mainmenu.gui", true, false, true);
@@ -3280,23 +2969,6 @@ int idSessionLocal::GetLocalClientNum()
 	else
 	{
 		return 0;
-	}
-}
-
-/*
-===============
-idSessionLocal::SetPlayingSoundWorld
-===============
-*/
-void idSessionLocal::SetPlayingSoundWorld()
-{
-	if (guiActive && (guiActive == guiMainMenu || guiActive == guiIntro || guiActive == guiLoading || (guiActive == guiMsg && !mapSpawned)))
-	{
-		soundSystem->SetPlayingSoundWorld(menuSoundWorld);
-	}
-	else
-	{
-		soundSystem->SetPlayingSoundWorld(sw);
 	}
 }
