+++ src/sound/snd_local.h
@@ -465,7 +465,7 @@ public:
 	virtual idSoundWorld *GetPlayingSoundWorld();
 
 	// sends the current playing sound world information to the sound hardware
-	virtual void Render();
+	virtual void Update();
 
 	// Mutes the SSG_MUSIC group
 	virtual void MuteBackgroundMusic(bool mute)
