+++ src/sound/sound.h
@@ -282,7 +282,7 @@ public:
 	virtual idSoundWorld *GetPlayingSoundWorld() = 0;
 
 	// Sends the current playing sound world information to the sound hardware.
-	virtual void Render() = 0;
+	virtual void Update() = 0;
 
 	virtual void MuteBackgroundMusic(bool mute) = 0;
